# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SessionRequest {
  username: string;
  fingerprint: string;
}

export interface SessionInfo {
  sessionId: string;
  username: string;
  createdAt: string;
  trialEndsAt: string;
  planType?: string | null;
  planExpiresAt?: string | null;
  isTrialActive: boolean;
  isPlanActive: boolean;
}

export interface GenerateImageRequest {
  sessionId: string;
  prompt: string;
  aesthetic: string;
  referenceImageBase64?: string | null;
}

export interface GeneratedImage {
  url: string;
  label: string;
  dimensions: string;
}

export interface GenerateImageResponse {
  images: GeneratedImage[];
  generationId: string;
}

export interface GenerateVideoRequest {
  sessionId: string;
  prompt: string;
  style: string;
  sourceImageUrl?: string | null;
}

export interface GenerateVideoResponse {
  videoUrl: string;
  generationId: string;
}

export interface GenerationHistoryItem {
  id: string;
  type: string;
  prompt: string;
  aesthetic: string;
  createdAt: string;
}

export interface GenerationHistoryResponse {
  items: GenerationHistoryItem[];
}

export type CreatePaymentRequestPlan =
  (typeof CreatePaymentRequestPlan)[keyof typeof CreatePaymentRequestPlan];

export const CreatePaymentRequestPlan = {
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface CreatePaymentRequest {
  sessionId: string;
  plan: CreatePaymentRequestPlan;
  payoneerTransactionId: string;
  amount: number;
}

export interface PaymentRecord {
  id: string;
  plan: string;
  amount: number;
  status: string;
  createdAt: string;
  expiresAt: string;
}

export interface PaymentHistoryResponse {
  payments: PaymentRecord[];
}

export interface PlanStatus {
  isActive: boolean;
  planType?: string | null;
  expiresAt?: string | null;
}