# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create or get existing user session
 */
export const CreateOrGetSessionBody = zod.object({
  username: zod.string(),
  fingerprint: zod.string(),
});

export const CreateOrGetSessionResponse = zod.object({
  sessionId: zod.string(),
  username: zod.string(),
  createdAt: zod.string(),
  trialEndsAt: zod.string(),
  planType: zod.string().nullish(),
  planExpiresAt: zod.string().nullish(),
  isTrialActive: zod.boolean(),
  isPlanActive: zod.boolean(),
});

/**
 * @summary Get session by ID
 */
export const GetSessionParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetSessionResponse = zod.object({
  sessionId: zod.string(),
  username: zod.string(),
  createdAt: zod.string(),
  trialEndsAt: zod.string(),
  planType: zod.string().nullish(),
  planExpiresAt: zod.string().nullish(),
  isTrialActive: zod.boolean(),
  isPlanActive: zod.boolean(),
});

/**
 * @summary Generate stream pack images
 */
export const GenerateImageBody = zod.object({
  sessionId: zod.string(),
  prompt: zod.string(),
  aesthetic: zod.string(),
  referenceImageBase64: zod.string().nullish(),
});

export const GenerateImageResponse = zod.object({
  images: zod.array(
    zod.object({
      url: zod.string(),
      label: zod.string(),
      dimensions: zod.string(),
    }),
  ),
  generationId: zod.string(),
});

/**
 * @summary Generate animated video
 */
export const GenerateVideoBody = zod.object({
  sessionId: zod.string(),
  prompt: zod.string(),
  style: zod.string(),
  sourceImageUrl: zod.string().nullish(),
});

export const GenerateVideoResponse = zod.object({
  videoUrl: zod.string(),
  generationId: zod.string(),
});

/**
 * @summary Get generation history
 */
export const GetGenerationHistoryParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetGenerationHistoryResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      type: zod.string(),
      prompt: zod.string(),
      aesthetic: zod.string(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Record a Payoneer payment
 */
export const CreatePaymentBody = zod.object({
  sessionId: zod.string(),
  plan: zod.enum(["weekly", "monthly"]),
  payoneerTransactionId: zod.string(),
  amount: zod.number(),
});

export const CreatePaymentResponse = zod.object({
  id: zod.string(),
  plan: zod.string(),
  amount: zod.number(),
  status: zod.string(),
  createdAt: zod.string(),
  expiresAt: zod.string(),
});

/**
 * @summary Get payment history for session
 */
export const GetPaymentHistoryParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetPaymentHistoryResponse = zod.object({
  payments: zod.array(
    zod.object({
      id: zod.string(),
      plan: zod.string(),
      amount: zod.number(),
      status: zod.string(),
      createdAt: zod.string(),
      expiresAt: zod.string(),
    }),
  ),
});

/**
 * @summary Verify active subscription plan
 */
export const VerifyPlanParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const VerifyPlanResponse = zod.object({
  isActive: zod.boolean(),
  planType: zod.string().nullish(),
  expiresAt: zod.string().nullish(),
});