# lib/api-zod/src/generated/types/createPaymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePaymentRequestPlan } from "./createPaymentRequestPlan";

export interface CreatePaymentRequest {
  sessionId: string;
  plan: CreatePaymentRequestPlan;
  payoneerTransactionId: string;
  amount: number;
}