# lib/api-zod/src/generated/types/generateImageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateImageRequest {
  sessionId: string;
  prompt: string;
  aesthetic: string;
  referenceImageBase64?: string | null;
}