# lib/api-zod/src/generated/types/generateImageResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedImage } from "./generatedImage";

export interface GenerateImageResponse {
  images: GeneratedImage[];
  generationId: string;
}