# lib/api-zod/src/generated/types/generateVideoResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateVideoResponse {
  videoUrl: string;
  generationId: string;
}