# lib/api-zod/src/generated/types/generatedImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedImage {
  url: string;
  label: string;
  dimensions: string;
}