# lib/api-zod/src/generated/types/generationHistoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerationHistoryItem {
  id: string;
  type: string;
  prompt: string;
  aesthetic: string;
  createdAt: string;
}