# lib/api-zod/src/generated/types/generationHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerationHistoryItem } from "./generationHistoryItem";

export interface GenerationHistoryResponse {
  items: GenerationHistoryItem[];
}