# lib/api-zod/src/generated/types/paymentHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SPARK GEN API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRecord } from "./paymentRecord";

export interface PaymentHistoryResponse {
  payments: PaymentRecord[];
}